# src/main/resources/static/swagger-ui/openapi3.yaml
openapi: 3.0.1
info:
  title: DevLog API
  description: DevLog API Documentation
  version: 0.1.0
servers:
- url: http://localhost:8080
...
paths:
...
  /api/tasks/{taskId}:
    get:
      tags:
      - Task
      summary: 작업 상세 조회 API
      description: 작업 상세 조회 API
      operationId: 작업 상세 조회 성공
      parameters:
      - name: taskId
        in: path
        description: 조회할 작업 고유 식별 ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
              examples:
                작업 상세 조회 성공:
                  value: "{\"status\":\"SUCCESS\",\"body\":{\"taskId\":1,\"title\"\
                    :\"인증 기능 구현\",\"description\":\"스프링 시큐리티 + JWT 기반 로그인 API 개발\"\
                    ,\"status\":\"TODO\",\"priority\":\"HIGH\",\"dueDate\":\"2026-06-12\"\
                    ,\"assigneeName\":\"test\",\"tags\":[{\"tagId\":10,\"name\":\"\
                    Backend\",\"color\":\"#FF5733\"}]},\"timestamp\":\"2026-06-05T16:25:34.2960048\"\
                    }"
    patch:
      tags:
      - Task
      summary: 작업 기본정보 수정 API
      description: 작업 기본정보 수정 API
      operationId: 작업 기본정보 수정 성공
      parameters:
      - name: taskId
        in: path
        description: 조회할 작업 고유 식별 ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
            examples:
              작업 기본정보 수정 성공:
                value: "{\"title\":\"수정된 제목\",\"description\":\"수정된 설명\"}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskIdResponse"
              examples:
                작업 기본정보 수정 성공:
                  value: "{\"status\":\"SUCCESS\",\"body\":{\"taskId\":1},\"timestamp\"\
                    :\"2026-06-05T16:25:33.9072317\"}"
...
  /api/projects/{projectId}/tasks:
    get:
      tags:
      - Task
      summary: 검색 조건 및 페이지네이션을 포함한 작업 목록 조회 API
      description: 검색 조건 및 페이지네이션을 포함한 작업 목록 조회 API
      operationId: 작업 목록 조회 성공
      parameters:
      - name: projectId
        in: path
        description: 조회할 프로젝트의 고유 식별 ID
        required: true
        schema:
          type: string
      - name: title
        in: query
        description: 검색할 작업 제목 키워드
        required: false
        schema:
          type: string
      - name: assigneeId
        in: query
        description: 담당자 고유 ID (전체 조회시 0 또는 NULL)
        required: false
        schema:
          type: string
      - name: status
        in: query
        description: "작업 상태 필터 (ALL, TODO, IN_PROGRESS, DONE, HOLD)"
        required: false
        schema:
          type: string
      - name: priority
        in: query
        description: "우선순위 필터 (ALL, LOW, MEDIUM, HIGH, URGENT)"
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: "조회할 페이지 번호 (0부터 시작, 기본값: 0)"
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: "한 페이지당 보여줄 데이터 개수 (기본값: 10)"
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: "정렬 기준 필드 및 방향 (기본값: dueDate,asc)"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
              examples:
                작업 목록 조회 성공:
                  value: "{\"status\":\"SUCCESS\",\"body\":{\"content\":[{\"taskId\"\
                    :1,\"title\":\"인증 기능 구현\",\"assigneeName\":\"개발자A\",\"status\"\
                    :\"TODO\",\"priority\":\"HIGH\",\"dueDate\":\"2026-06-12\"}],\"\
                    pageInfo\":{\"currentPage\":1,\"pageSize\":10,\"totalElements\"\
                    :1,\"totalPages\":1,\"isFirst\":true,\"isLast\":true}},\"timestamp\"\
                    :\"2026-06-05T16:25:34.4275273\"}"
    post:
      tags:
      - Task
      summary: 작업(Task) 생성 API
      description: 작업(Task) 생성 API
      operationId: 작업 생성 성공
      parameters:
      - name: projectId
        in: path
        description: 조회할 프로젝트의 고유 식별 ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
            examples:
              작업 생성 성공:
                value: "{\"title\":\"인증 기능 구현\",\"description\":\"스프링 시큐리티 + JWT 기\
                  반 로그인 API 개발\",\"status\":\"TODO\",\"priority\":\"HIGH\",\"dueDate\"\
                  :\"2026-06-12\",\"tagIds\":[1],\"assigneeId\":1}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskIdResponse"
              examples:
                작업 생성 성공:
                  value: "{\"status\":\"SUCCESS\",\"body\":{\"taskId\":1},\"timestamp\"\
                    :\"2026-06-05T16:25:34.737144\"}"
  /api/tasks/{taskId}/assignee:
    patch:
      tags:
      - Task
      summary: 작업 담당자 변경 API
      description: 작업 담당자 변경 API
      operationId: 작업 담당자 변경 성공
      parameters:
      - name: taskId
        in: path
        description: 조회할 작업 고유 식별 ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/ChangeAssigneeRequest"
            examples:
              작업 담당자 변경 성공:
                value: "{\"userId\":2}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskIdResponse"
              examples:
                작업 담당자 변경 성공:
                  value: "{\"status\":\"SUCCESS\",\"body\":{\"taskId\":1},\"timestamp\"\
                    :\"2026-06-05T16:25:32.119297\"}"
  /api/tasks/{taskId}/priority:
    patch:
      tags:
      - Task
      summary: 작업 우선순위 변경 API
      description: 작업 우선순위 변경 API
      operationId: 작업 우선순위 변경 성공
      parameters:
      - name: taskId
        in: path
        description: 조회할 작업 고유 식별 ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/ChangePriorityRequest"
            examples:
              작업 우선순위 변경 성공:
                value: "{\"priority\":\"MEDIUM\"}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskIdResponse"
              examples:
                작업 우선순위 변경 성공:
                  value: "{\"status\":\"SUCCESS\",\"body\":{\"taskId\":1},\"timestamp\"\
                    :\"2026-06-05T16:25:33.2867648\"}"
  /api/tasks/{taskId}/status:
    patch:
      tags:
      - Task
      summary: 작업 상태 변경 API
      description: 작업 상태 변경 API
      operationId: 작업 상태 변경 성공
      parameters:
      - name: taskId
        in: path
        description: 조회할 작업 고유 식별 ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/ChangeStatusRequest"
            examples:
              작업 상태 변경 성공:
                value: "{\"status\":\"IN_PROGRESS\"}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskIdResponse"
              examples:
                작업 상태 변경 성공:
                  value: "{\"status\":\"SUCCESS\",\"body\":{\"taskId\":1},\"timestamp\"\
                    :\"2026-06-05T16:25:33.8530524\"}"
  /api/projects/{projectId}/tasks/kanban:
    get:
      tags:
      - Task
      summary: 프로젝트별 칸반보드 작업 목록 조회 API
      description: 프로젝트별 칸반보드 작업 목록 조회 API
      operationId: 칸반보드 작업 목록 조회 성공
      parameters:
      - name: projectId
        in: path
        description: 조회할 프로젝트의 고유 식별 ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KanbanBoardResponse"
              examples:
                칸반보드 작업 목록 조회 성공:
                  value: "{\"status\":\"SUCCESS\",\"body\":{\"board\":{\"TODO\":[{\"\
                    taskId\":1,\"title\":\"인증 기능 구현\",\"description\":\"스프링 시큐리티 +\
                    \ JWT 기반 로그인 API 개발\",\"status\":\"TODO\",\"priority\":\"HIGH\"\
                    ,\"dueDate\":\"2026-06-12\",\"assigneeName\":\"데브로그팀원\",\"tags\"\
                    :[{\"tagId\":10,\"name\":\"Frontend\",\"color\":\"#3357FF\"}]}],\"\
                    IN_PROGRESS\":[],\"DONE\":[],\"HOLD\":[]}},\"timestamp\":\"2026-06-05T16:25:34.5037309\"\
                    }"
components:
  schemas:
...
    TaskResponse:
      title: TaskResponse
      required:
      - status
      - timestamp
      type: object
      properties:
        body:
          required:
          - dueDate
          - priority
          - status
          - tags
          - taskId
          - title
          type: object
          properties:
            assigneeName:
              type: string
              description: 담당자 이름
              nullable: true
            dueDate:
              type: string
              description: 마감일 (YYYY-MM-DD)
            description:
              type: string
              description: 작업 상세 설명
              nullable: true
            priority:
              type: string
              description: "우선순위 (HIGH, MEDIUM, LOW)"
            title:
              type: string
              description: 작업 제목
            taskId:
              type: number
              description: 작업 고유 ID
            tags:
              type: array
              description: 작업 태그 목록
              items:
                required:
                - color
                - name
                - tagId
                type: object
                properties:
                  color:
                    type: string
                    description: 태그 색상 헥사 코드
                  tagId:
                    type: number
                    description: 태그 고유 ID
                  name:
                    type: string
                    description: 태그 이름
            status:
              type: string
              description: 작업 현재 상태
        timestamp:
          type: string
          description: 응답 발행 일시
        status:
          type: string
          description: 응답 상태 코드/메시지
...
    KanbanBoardResponse:
      title: KanbanBoardResponse
      required:
      - status
      - timestamp
      type: object
      properties:
        body:
          required:
          - board
          type: object
          properties:
            board:
              required:
              - DONE
              - HOLD
              - IN_PROGRESS
              - TODO
              type: object
              properties:
                IN_PROGRESS:
                  type: array
                  description: 진행 중(IN_PROGRESS) 작업 목록
                  items:
                    oneOf:
                    - type: object
                    - type: boolean
                    - type: string
                    - type: number
                TODO:
                  type: array
                  description: 할 일(TODO) 작업 목록
                  items:
                    oneOf:
                    - type: object
                    - type: boolean
                    - type: string
                    - type: number
                DONE:
                  type: array
                  description: 완료(DONE) 작업 목록
                  items:
                    oneOf:
                    - type: object
                    - type: boolean
                    - type: string
                    - type: number
                '*':
                  type: array
                  items:
                    required:
                    - dueDate
                    - priority
                    - status
                    - tags
                    - taskId
                    - title
                    type: object
                    properties:
                      assigneeName:
                        type: string
                        description: 담당자 이름
                        nullable: true
                      dueDate:
                        type: string
                        description: 마감일 (YYYY-MM-DD)
                      description:
                        type: string
                        description: 작업 상세 설명
                        nullable: true
                      priority:
                        type: string
                        description: "우선순위 (HIGH, MEDIUM, LOW)"
                      title:
                        type: string
                        description: 작업 제목
                      taskId:
                        type: number
                        description: 작업 고유 ID
                      tags:
                        type: array
                        description: 작업 태그 목록
                        items:
                          required:
                          - color
                          - name
                          - tagId
                          type: object
                          properties:
                            color:
                              type: string
                              description: 태그 색상 헥사 코드
                            tagId:
                              type: number
                              description: 태그 고유 ID
                            name:
                              type: string
                              description: 태그 이름
                      status:
                        type: string
                        description: 작업 현재 상태
                HOLD:
                  type: array
                  description: 보류(HOLD) 작업 목록
                  items:
                    oneOf:
                    - type: object
                    - type: boolean
                    - type: string
                    - type: number
              description: 상태별 칸반보드 데이터 Map
        timestamp:
          type: string
          description: 응답 발행 일시
        status:
          type: string
          description: 응답 상태 코드/메시지
...
    CreateTaskRequest:
      title: CreateTaskRequest
      required:
      - assigneeId
      - dueDate
      - priority
      - status
      - title
      type: object
      properties:
        tagIds:
          type: array
          description: 연결할 태그 식별 ID 리스트
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        dueDate:
          type: string
          description: 마감일 (YYYY-MM-DD)
        description:
          type: string
          description: 작업 상세 설명
          nullable: true
        priority:
          type: string
          description: "우선순위 (HIGH, MEDIUM, LOW 등)"
        title:
          type: string
          description: 작업 제목
        assigneeId:
          type: number
          description: 담당 사용자 식별 ID
        status:
          type: string
          description: "작업 상태 (TODO, IN_PROGRESS, DONE 등)"
...
    TaskIdResponse:
      title: TaskIdResponse
      required:
      - status
      - timestamp
      type: object
      properties:
        body:
          required:
          - taskId
          type: object
          properties:
            taskId:
              type: number
              description: 생성된 작업 고유 ID
        timestamp:
          type: string
          description: 응답 발행 일시
        status:
          type: string
          description: 응답 상태 코드/메시지
    UpdateTaskRequest:
      title: UpdateTaskRequest
      required:
      - title
      type: object
      properties:
        description:
          type: string
          description: 수정된 설명
          nullable: true
        title:
          type: string
          description: 수정된 제목
...
    ChangePriorityRequest:
      title: ChangePriorityRequest
      required:
      - priority
      type: object
      properties:
        priority:
          type: string
          description: 변경된 우선순위 값
    ChangeAssigneeRequest:
      title: ChangeAssigneeRequest
      required:
      - userId
      type: object
      properties:
        userId:
          type: number
          description: 변경된 담당자 고유 식별 ID
...
    ChangeStatusRequest:
      title: ChangeStatusRequest
      required:
      - status
      type: object
      properties:
        status:
          type: string
          description: 변경된 상태값
...
    TaskListResponse:
      title: TaskListResponse
      required:
      - body
      - status
      - timestamp
      type: object
      properties:
        body:
          required:
          - content
          - pageInfo
          type: object
          properties:
            pageInfo:
              required:
              - currentPage
              - isFirst
              - isLast
              - pageSize
              - totalElements
              - totalPages
              type: object
              properties:
                isFirst:
                  type: boolean
                  description: 첫 페이지 여부
                isLast:
                  type: boolean
                  description: 마지막 페이지 여부
                totalPages:
                  type: number
                  description: 총 페이지 수
                pageSize:
                  type: number
                  description: 페이지 당 노출 데이터 개수
                currentPage:
                  type: number
                  description: 현재 페이지 번호 (1-indexed)
                totalElements:
                  type: number
                  description: 총 데이터 개수
              description: 페이징 메타데이터
            content:
              type: array
              description: 조회된 작업 목록 배열
              items:
                required:
                - dueDate
                - priority
                - status
                - taskId
                - title
                type: object
                properties:
                  assigneeName:
                    type: string
                    description: 담당자 이름
                    nullable: true
                  dueDate:
                    type: string
                    description: 마감일 (YYYY-MM-DD)
                  priority:
                    type: string
                    description: 작업 우선순위
                  title:
                    type: string
                    description: 작업 제목
                  taskId:
                    type: number
                    description: 작업 고유 ID
                  status:
                    type: string
                    description: 작업 상태
          description: 응답 본문
        timestamp:
          type: string
          description: 응답 발행 일시
        status:
          type: string
          description: 응답 상태 코드/메시지